+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Fixed**: better printing in scroll view
 - **Fixed**: better printing in preview
 - **Changed**: design
+- **Changed**: scale only to 1 in scroll view
 - **Changed**: few more things...
 
 ## 0.4.8 : 2023-10-05